--- scrap-app/src/Login.css
@@ -1,18 +1,0 @@
-.logIn{
-  display: flex;
-  justify-content: center;
-  align-items: center; 
-  height: 100vh; 
-}
-
-.container{
-  border-radius: 20px;
-  background-color: #d9d9d9;
-  width: 400px;
-  height: 500px; 
-}
-
-.inputContainer{
-  margin-top: 1%;
-  margin-bottom: 1%;
-}
