+++ scrap-app/src/components/Login/index.jsx
@@ -1,0 +1,75 @@
+import { useState } from "react";
+import axios from "axios";
+import { useNavigate } from "react-router-dom";
+import styles from "./styles.module.css";
+import { Link } from "react-router-dom";
+
+const Login = ({ setUser }) => {
+    const [data, setData] = useState({ username: "", password: "" });
+    const [error, setError] = useState("");
+    const navigate = useNavigate();
+
+    const handleChange = ({ currentTarget: input }) => {
+        setData({ ...data, [input.name]: input.value });
+    };
+
+  
+    const handleSubmit = async (e) => {
+      e.preventDefault();
+      try {
+        const url = "http://localhost:4000/api/auth";
+        const { data: res } = await axios.post(url, data);
+        localStorage.setItem("token", res.data);
+        setUser(true); // Update user state
+        navigate("/"); // Redirect to the home page
+      } catch (error) {
+        if (error.response?.status >= 400 && error.response?.status <= 500) {
+          setError(error.response.data.message);
+        }
+      }
+    };
+  
+    return (
+      <div className={styles.login_container}>
+        <div className={styles.login_form_container}>
+                <div className={styles.left}>
+                    <form className={styles.form_container} onSubmit={handleSubmit}>
+                        <h1>Login to Your Account</h1>
+                        <input
+                            type="username"
+                            placeholder="Username"
+                            name="username"
+                            onChange={handleChange}
+                            value={data.username}
+                            required
+                            className={styles.input}
+                        />
+                        <input
+                            type="password"
+                            placeholder="Password"
+                            name="password"
+                            onChange={handleChange}
+                            value={data.password}
+                            required
+                            className={styles.input}
+                        />
+                        {error && <div className={styles.error_msg}>{error}</div>}
+                        <button type="submit" className={styles.green_btn}>
+                            Sign In
+                        </button>
+                    </form>
+                </div>
+                <div className={styles.right}>
+                    <h1>New Here?</h1>
+                    <Link to="/signup">
+                        <button type="button" className={styles.white_btn}>
+                            Sign Up
+                        </button>
+                    </Link>
+                </div>
+            </div>
+      </div>
+    );
+  };
+  
+  export default Login;
