--- scrap-app/src/components/Main/index.jsx
@@ -1,21 +1,0 @@
-import styles from "./styles.module.css";
-
-const Main = () => {
-	const handleLogout = () => {
-		localStorage.removeItem("token");
-		window.location.reload();
-	};
-
-	return (
-		<div className={styles.main_container}>
-			<nav className={styles.navbar}>
-				<h1>scrap</h1>
-				<button className={styles.white_btn} onClick={handleLogout}>
-					Logout
-				</button>
-			</nav>
-		</div>
-	);
-};
-
-export default Main;
